fixed to restaurant id

+++ ORM/entities/TblThirdparty.ts
@@ -2,7 +2,7 @@ import {Column, Entity, OneToMany, PrimaryGeneratedColumn} from "typeorm";
 
 @Entity()
 export class TblThirdparty {
-    @PrimaryGeneratedColumn({type: "bigint", name: "id"})
+    @PrimaryGeneratedColumn({type: "int", name: "id"})
     id : number;
 
     @Column("varchar", {
